+++ misc/mapFile.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-void* mapFile_map(const char* file_name, uint64_t* size);
+void* mapFile_map(const char* file_name, size_t* size);
 
 #endif
